Move BackupProgress* files out of fdbserver/logsystem
