+++ fdbserver/clustercontroller/ClusterRecovery.actor.cpp
@@ -24,7 +24,7 @@
 #include "fdbclient/FDBTypes.h"
 #include "fdbrpc/sim_validation.h"
 #include "fdbserver/logsystem/ApplyMetadataMutation.h"
-#include "fdbserver/logsystem/BackupProgress.h"
+#include "fdbserver/core/BackupProgress.h"
 #include "ClusterRecovery.actor.h"
 #include "fdbserver/core/Knobs.h"
 #include "fdbserver/core/MasterInterface.h"
