+++ fdbserver/core/BackupProgress.cpp
@@ -18,7 +18,9 @@
  * limitations under the License.
  */
 
-#include "fdbserver/logsystem/BackupProgress.h"
+#include <algorithm>
+
+#include "fdbserver/core/BackupProgress.h"
 
 #include "fdbclient/NativeAPI.actor.h"
 #include "fdbclient/SystemData.h"
@@ -50,8 +52,6 @@ void BackupProgress::updateTagVersions(std::map<Tag, Version>* tagVersions,
                                        Version adjustedBeginVersion,
                                        LogEpoch epoch) {
 	for (const auto& [tag, savedVersion] : progress) {
-		// If tag is not in "tags", it means the old epoch has more tags than
-		// new epoch's tags. Just ignore the tag here.
 		auto n = tags->erase(tag);
 		if (n > 0 && savedVersion < endVersion - 1) {
 			const Version beginVersion =
@@ -71,16 +71,13 @@ std::map<std::tuple<LogEpoch, Version, int>, std::map<Tag, Version>> BackupProgr
 	std::map<std::tuple<LogEpoch, Version, int>, std::map<Tag, Version>> toRecruit;
 
 	if (!backupStartedValue.present())
-		return toRecruit; // No active backups
+		return toRecruit;
 
 	Version lastEnd = invalidVersion;
 	for (const auto& [epoch, info] : epochInfos) {
 		std::set<Tag> tags = enumerateLogRouterTags(info.logRouterTags);
 		std::map<Tag, Version> tagVersions;
 
-		// Sometimes, an epoch's begin version is lower than the previous epoch's
-		// end version. In this case, adjust the epoch's begin version to be the
-		// same as previous end version.
 		Version adjustedBeginVersion = lastEnd > info.epochBegin ? lastEnd : info.epochBegin;
 		lastEnd = info.epochEnd;
 
@@ -89,7 +86,6 @@ std::map<std::tuple<LogEpoch, Version, int>, std::map<Tag, Version>> BackupProgr
 			std::set<Tag> toCheck = tags;
 			for (auto current = progressIt; current != progress.begin() && !toCheck.empty();) {
 				auto prev = std::prev(current);
-				// Previous epoch is gone, consolidate the progress.
 				for (auto [tag, version] : prev->second) {
 					if (toCheck.contains(tag)) {
 						progressIt->second[tag] = std::max(version, progressIt->second[tag]);
@@ -105,9 +101,6 @@ std::map<std::tuple<LogEpoch, Version, int>, std::map<Tag, Version>> BackupProgr
 			    progress.rend(),
 			    [epoch = epoch](const std::pair<LogEpoch, std::map<Tag, Version>>& p) { return p.first < epoch; });
 			while (!(rit == progress.rend())) {
-				// A partial recovery can result in empty epoch that copies previous
-				// epoch's version range. In this case, we should check previous
-				// epoch's savedVersion.
 				int savedMore = 0;
 				for (auto [tag, version] : rit->second) {
 					if (version >= info.epochBegin) {
@@ -115,9 +108,6 @@ std::map<std::tuple<LogEpoch, Version, int>, std::map<Tag, Version>> BackupProgr
 					}
 				}
 				if (savedMore > 0) {
-					// The logRouterTags are the same
-					// ASSERT(info.logRouterTags == epochTags[rit->first]);
-
 					updateTagVersions(&tagVersions, &tags, rit->second, info.epochEnd, adjustedBeginVersion, epoch);
 					if (tags.empty())
 						break;
@@ -126,7 +116,7 @@ std::map<std::tuple<LogEpoch, Version, int>, std::map<Tag, Version>> BackupProgr
 			}
 		}
 
-		for (const Tag& tag : tags) { // tags without progress data
+		for (const Tag& tag : tags) {
 			tagVersions.insert({ tag, adjustedBeginVersion });
 			TraceEvent("BackupVersionRange", dbgid)
 			    .detail("OldEpoch", epoch)
@@ -142,7 +132,6 @@ std::map<std::tuple<LogEpoch, Version, int>, std::map<Tag, Version>> BackupProgr
 	return toRecruit;
 }
 
-// Save each tag's savedVersion for all epochs into "bStatus".
 Future<Void> getBackupProgress(Database cx, UID dbgid, Reference<BackupProgress> bStatus, Severity severity) {
 	Transaction tr(cx);
